requirements.txt
# Neo4j and Graph Database
neo4j==5.15.0

# CrewAI Framework
crewai==0.74.0
crewai-tools>=0.13.1

# Chemistry and Molecular Analysis
rdkit==2023.9.4
chembl-structure-pipeline==1.2.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Machine Learning (optional, for enhanced predictions)
scikit-learn==1.3.2
xgboost==2.0.3

# LLM Integration
langchain>=0.2.16
langchain-openai>=0.0.2
openai>=1.6.1

# Utilities
python-dotenv==1.0.0
pydantic>=2.7.4
